requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
groq==0.12.0
pydantic==2.9.2
python-multipart==0.0.9
streamlit==1.38.0
requests==2.32.3
